webstack_django/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-cors-headers>=4.3.1
django-crispy-forms>=2.1  # Équivalent à Flask-WTF
crispy-bootstrap4>=2023.1
supabase>=1.0.3
requests>=2.31.0
python-jose>=3.3.0
django-allauth==0.60.1    # Équivalent à Flask-Login
django-environ==0.11.2    # Pour la gestion des variables d'environnement
djangorestframework>=3.14.0
Pillow>=10.1.0
django-storages>=1.14.2
django-redis>=5.4.0
pyjwt>=2.8.0
